backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-redis==5.4.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
requests==2.32.3
numpy==1.26.4
redis==5.0.8
